modular

+++ scripts/docker/setup.sh
@@ -10,6 +10,11 @@ sudo rm -rf $WORK_DIR
 
 printf $warning "\n\n------------ Unpacking Docker Release ------------\n\n"
 
+if [[ ! -f ~/release/takserver*.zip  ]]; then
+    printf $warning "\n\n------------ No TAK Server Package found in ~/release/ ------------\n\n"
+    exit
+fi
+
 unzip ~/release/takserver*.zip -d ~/
 mv ~/takserver* ${WORK_DIR}
 chown -R $USER:$USER ${WORK_DIR}
